+++ example/src/Examples/Video/Stories/index.ts
@@ -1,0 +1,1 @@
+export * from "./Stories";
