--- example/src/Examples/Video/Transitions/Math.ts
@@ -1,11 +1,0 @@
-export const snapPoint = (
-  value: number,
-  velocity: number,
-  points: ReadonlyArray<number>
-): number => {
-  "worklet";
-  const point = value + 0.2 * velocity;
-  const deltas = points.map((p) => Math.abs(point - p));
-  const minDelta = Math.min.apply(null, deltas);
-  return points.filter((p) => Math.abs(point - p) === minDelta)[0];
-};
