+++ example/src/Examples/Video/gltransitions/glitchMemories.ts
@@ -1,5 +1,5 @@
 /* eslint-disable max-len */
-import { glsl } from "../../../../components/ShaderLib";
+import { glsl } from "../../../components/ShaderLib";
 
 import type { Transition } from "./Base";
