+++ example/src/Examples/Video/gltransitions/pageCurl.ts
@@ -1,6 +1,6 @@
 /* eslint-disable max-len */
 
-import { glsl } from "../../../../components/ShaderLib";
+import { glsl } from "../../../components/ShaderLib";
 
 import type { Transition } from "./Base";
