+++ example/src/Examples/Video/gltransitions/swirl.ts
@@ -1,4 +1,4 @@
-import { glsl } from "../../../../components/ShaderLib";
+import { glsl } from "../../../components/ShaderLib";
 
 import type { Transition } from "./Base";
