+++ example/src/components/Animations.tsx
@@ -40,3 +40,15 @@ export const translate = ({
   "worklet";
   return [{ translateX: x }, { translateY: y }];
 };
+
+export const snapPoint = (
+  value: number,
+  velocity: number,
+  points: ReadonlyArray<number>
+): number => {
+  "worklet";
+  const point = value + 0.2 * velocity;
+  const deltas = points.map((p) => Math.abs(point - p));
+  const minDelta = Math.min.apply(null, deltas);
+  return points.filter((p) => Math.abs(point - p) === minDelta)[0];
+};
